Ensure workflow is updated for snapshot when we CRUD a node.
